BookmarksVC cleanup: remove unused code
